+++ core/smd-api/src/test/java/app/dodb/smd/api/command/AnnotatedCommandHandlerTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 
 import java.time.Instant;
 
+import static app.dodb.smd.api.utils.LoggingUtils.logClass;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -31,7 +32,7 @@ class AnnotatedCommandHandlerTest {
     void handle_withoutCommandParameter() {
         assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithoutCommandParameter()))
             .isInstanceOf(IllegalArgumentException.class)
-            .hasMessageStartingWith("Invalid command handler: method must include a parameter of type Command.");
+            .hasMessageStartingWith("Invalid handler: method must include a parameter of type %s.".formatted(logClass(Command.class)));
     }
 
     @Test
@@ -49,31 +50,66 @@ class AnnotatedCommandHandlerTest {
     }
 
     @Test
+    void handle_withCommandAndMultipleMetadataValueParameters() {
+        var registry = AnnotatedCommandHandler.from(new CommandHandlerWithCommandAndMultipleMetadataValueParameters());
+
+        assertThat(registry.commandHandlers()).hasSize(1);
+    }
+
+    @Test
     void handle_withoutMetadataValueAnnotation() {
         assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithoutMetadataValueAnnotation()))
             .isInstanceOf(IllegalArgumentException.class)
-            .hasMessageStartingWith("Invalid command handler: metadata value parameter must be annotated with @MetadataValue.");
+            .hasMessageStartingWith("Invalid handler: metadata value parameter must be annotated with @MetadataValue.");
     }
 
     @Test
     void handle_withIncorrectMetadataValueAnnotation() {
         assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithIncorrectMetadataValueAnnotation()))
             .isInstanceOf(IllegalArgumentException.class)
-            .hasMessageStartingWith("Invalid command handler: only parameters of type String can be annotated with @MetadataValue.");
+            .hasMessageStartingWith("Invalid handler: only parameters of type String can be annotated with @MetadataValue.");
+    }
+
+    @Test
+    void handle_withDuplicateMessageIdParameters() {
+        assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithDuplicateMessageIdParameters()))
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessageStartingWith("Invalid handler: method must only include one parameter of type %s.".formatted(logClass(MessageId.class)));
+    }
+
+    @Test
+    void handle_withDuplicateMetadataParameters() {
+        assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithDuplicateMetadataParameters()))
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessageStartingWith("Invalid handler: method must only include one parameter of type %s.".formatted(logClass(Metadata.class)));
+    }
+
+    @Test
+    void handle_withDuplicatePrincipalParameters() {
+        assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithDuplicatePrincipalParameters()))
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessageStartingWith("Invalid handler: method must only include one parameter of type %s.".formatted(logClass(app.dodb.smd.api.metadata.principal.Principal.class)));
+    }
+
+    @Test
+    void handle_withDuplicateInstantParameters() {
+        assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithDuplicateInstantParameters()))
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessageStartingWith("Invalid handler: method must only include one parameter of type %s.".formatted(logClass(Instant.class)));
     }
 
     @Test
     void handle_withoutParameters() {
         assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithoutParameters()))
             .isInstanceOf(IllegalArgumentException.class)
-            .hasMessageStartingWith("Invalid command handler: method must have at least one parameter.");
+            .hasMessageStartingWith("Invalid handler: method must have at least one parameter.");
     }
 
     @Test
     void handle_withMultipleCommandTypes() {
         assertThatThrownBy(() -> AnnotatedCommandHandler.from(new CommandHandlerWithMultipleCommandTypes()))
             .isInstanceOf(IllegalArgumentException.class)
-            .hasMessageStartingWith("Invalid command handler: method must only include one Command as a parameter.");
+            .hasMessageStartingWith("Invalid handler: method must only include one parameter of type %s.".formatted(logClass(Command.class)));
     }
 
     @Test
@@ -148,6 +184,14 @@ class AnnotatedCommandHandlerTest {
         }
     }
 
+    public static class CommandHandlerWithCommandAndMultipleMetadataValueParameters {
+
+        @CommandHandler
+        public String handle(CommandForTest command, @MetadataValue("value1") String value1, @MetadataValue("value2") String value2) {
+            return "";
+        }
+    }
+
     public static class CommandHandlerWithoutMetadataValueAnnotation {
 
         @CommandHandler
@@ -160,6 +204,38 @@ class AnnotatedCommandHandlerTest {
 
         @CommandHandler
         public String handle(CommandForTest command, @MetadataValue("property") Metadata metadata) {
+            return "";
+        }
+    }
+
+    public static class CommandHandlerWithDuplicateMessageIdParameters {
+
+        @CommandHandler
+        public String handle(CommandForTest command, MessageId messageId1, MessageId messageId2) {
+            return "";
+        }
+    }
+
+    public static class CommandHandlerWithDuplicateMetadataParameters {
+
+        @CommandHandler
+        public String handle(CommandForTest command, Metadata metadata1, Metadata metadata2) {
+            return "";
+        }
+    }
+
+    public static class CommandHandlerWithDuplicatePrincipalParameters {
+
+        @CommandHandler
+        public String handle(CommandForTest command, SimplePrincipal principal1, SimplePrincipal principal2) {
+            return "";
+        }
+    }
+
+    public static class CommandHandlerWithDuplicateInstantParameters {
+
+        @CommandHandler
+        public String handle(CommandForTest command, Instant instant1, Instant instant2) {
             return "";
         }
     }